+++ Numeric/GaussQuadraturePri.cpp
@@ -14,7 +14,8 @@ IntPt *getGQPriPts(int order)
   int nLin = (order + 3) / 2;
   int nTri = getNGQTPts(order);
   int n = nLin * nTri;
-  if(GQP.size() < order + 1) GQP.resize(order + 1, nullptr);
+  if(static_cast<int>(GQP.size()) < order + 1)
+    GQP.resize(order + 1, nullptr);
   if(!GQP[order]) {
     double *linPt, *linWt;
     IntPt *triPts = getGQTPts(order);
